moved hyper-specific upgrade implementation into its own module
